Add binary search implementation
I believe method signature for binary search was written incorrectly.
The original signature made binary search itself Comparable, we instead
want to specify that the list given to binary search contain comparable
items.

I rewrote the signature and implemented binary search. None of the tests
had to be rewritten which suggests that the new generics specification
is in line with what the test writer intended.

The junit test order has been modified so that the tests run in the
order listed in the file. This makes it easier to check if all the tests
pertaining to a specific problem are passing.